rename file to match new terminology
